\name{refA}
\docType{data}
\alias{refA}
\title{ MAQC reference data for the Illumina platform}
\description{ A subset of the MAQC's 'A' RNA reference dataset }
\details{
The Microarray Quality Control Consortium (MAQC) has generated a reference 
dataset of Human-6 BeadChip 48K v1.0 from 100\% of Stratagene 
Universal Reference RNA. Three different test (\_1\_ to \_3\_) sites have processed 
this reference in 5 replicates (A1 to A5). 
}
\format{an object of type \code{"LumiBatch"}}
\seealso{refB, refC, refD}
\keyword{datasets}
