\name{refB}
\docType{data}
\alias{refB}
\title{ MAQC reference data for the Illumina platform}
\description{ A subset of the MAQC's 'B' RNA reference dataset }
\details{
The Microarray Quality Control Consortium (MAQC) has generated a reference 
dataset of Human-6 BeadChip 48K v1.0 from 100\% of Ambion Brain 
Reference RNA. Three different test (\_1\_ to \_3\_) sites have processed 
this reference in 5 replicates (B1 to B5). 
}
\format{an object of type \code{"LumiBatch"}}
\seealso{refA, refC, refD}
\keyword{datasets}
