\name{refC}
\docType{data}
\alias{refC}
\title{ MAQC reference data for the Illumina platform}
\description{ A subset of the MAQC's 'C' RNA reference dataset }
\details{
The Microarray Quality Control Consortium (MAQC) has generated a reference 
dataset of Human-6 BeadChip 48K v1.0 from 75% of Stratagene Universal Reference 
RNA and 25% Ambion Brain reference RNA. Three different test (\_1\_ to \_3\_) sites 
have processed this reference in 4 (site 1) or 5 replicates (C1 to C5). 
}
\format{an object of type \code{"LumiBatch"}}
\seealso{refA, refB, refD}
\keyword{datasets}
