\name{refD}
\docType{data}
\alias{refD}
\title{ MAQC reference data for the Illumina platform}
\description{ A subset of the MAQC's 'D' RNA reference dataset }
\details{
The Microarray Quality Control Consortium (MAQC) has generated a reference 
dataset of Human-6 BeadChip 48K v1.0 from 25% of Stratagene Universal Reference 
RNA and 75% Ambion Brain reference RNA. Three different test (\_1\_ to \_3\_) sites 
have processed this reference in 5 replicates (D1 to D5). 
}
\format{an object of type \code{"LumiBatch"}}
\seealso{refA, refB, refC}
\keyword{datasets}
